requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==19.8.1
Jinja2==2.10.3
MarkupSafe==1.0
python-dateutil==2.7.0
SQLAlchemy==1.2.11
webargs==5.5.2
Werkzeug==0.16.0
bcrypt==3.1.4
pyodbc==4.0.27
azure-functions==1.0.5
azf-wsgi==0.3.0
